single Character page is done and working

+++ javascript/singleCharacter.js
@@ -23,16 +23,19 @@ function displayData(data){
 }
 
 const buttonDelete=document.getElementById("buttonDeleteCharater");
-buttonDelete.addEventListener("click", deleteCharacter());
+buttonDelete.addEventListener("click", deleteCharacter);
 
 function deleteCharacter(){
-    console.log('func')
-    fetch("https://character-database.becode.xyz/characters/" +postName ,{method: 'DELETE'})
-    .then(response => {
-        if (response.ok){
-            console.log("DELETED")
-        }
-    })
+    if(confirm("Are you sure you want to delete this character?")){
+        fetch("https://character-database.becode.xyz/characters/" +postName ,{method: 'DELETE'})
+        .then(response => {
+            if (response.ok){
+                console.log("DELETED")
+            }
+        })
+    }
+    else{console.log("You did not delete this character")}
+    
 
 }
